@@ -47,7 +47,7 @@ namespace PaperMalKing.UpdatesProviders.Base.Features
 		{
 			if (!features.Any())
 				return;
-			this.Logger.LogInformation("Trying to enable {Features} feature for {Username}", features, context.Member.DisplayName);
+			this.Logger.LogInformation("Trying to enable {Features} feature for {Username}", features!, context.Member!.DisplayName);
 			try
 			{
 				await this.UserFeaturesService.EnableFeaturesAsync(features, context.User.Id).ConfigureAwait(false);
@@ -71,7 +71,7 @@ namespace PaperMalKing.UpdatesProviders.Base.Features
 		{
 			if (!features.Any())
 				return;
-			this.Logger.LogInformation("Trying to disable {Features} feature for {Username}", features, context.Member.DisplayName);
+			this.Logger.LogInformation("Trying to disable {Features} feature for {Username}", features!, context.Member!.DisplayName);
 			try
 			{
 				await this.UserFeaturesService.DisableFeaturesAsync(features, context.User.Id).ConfigureAwait(false);
