@@ -68,7 +68,7 @@ namespace PaperMalKing.Commands
 											.Replace("@here", "@\u200bhere", StringComparison.Ordinal),
 				Timestamp = DateTime.Now,
 				Color = DiscordColor.Blue
-			}.WithAuthor($"{context.Member.Username}#{context.Member.Discriminator}", iconUrl: context.Member.AvatarUrl);
+			}.WithAuthor($"{context.Member!.Username}#{context.Member.Discriminator}", iconUrl: context.Member!.AvatarUrl);
 			try
 			{
 				await channelToSayIn.SendMessageAsync(embed: embed).ConfigureAwait(false);
